Migrate DatePicker to Styled Components (#336)
